Refactor exceptions #12
- Refactor all exceptions in as3ninja.exceptions
- Update all modules accordingly
- Update tests accordingly
- Code formatting updates
