+++ tests/test_jinja2_j2ninja.py
@@ -4,14 +4,13 @@ from as3ninja.jinja2 import J2Ninja
 
 
 class Test_J2Ninja:
-
     @staticmethod
     def test_registerfunction():
         @J2Ninja.registerfunction
         def my_function():
             pass
 
-        assert J2Ninja.functions['my_function'] == my_function
+        assert J2Ninja.functions["my_function"] == my_function
 
     @staticmethod
     def test_registerfilter():
@@ -19,7 +18,7 @@ class Test_J2Ninja:
         def my_filter():
             pass
 
-        assert J2Ninja.filters['my_filter'] == my_filter
+        assert J2Ninja.filters["my_filter"] == my_filter
 
     @staticmethod
     def test_registertest():
@@ -27,7 +26,7 @@ class Test_J2Ninja:
         def my_test():
             return True
 
-        assert J2Ninja.tests['my_test'] == my_test
+        assert J2Ninja.tests["my_test"] == my_test
 
     @staticmethod
     def test_registerfilterfunction():
@@ -36,5 +35,5 @@ class Test_J2Ninja:
         def my_filterfunction():
             return True
 
-        assert J2Ninja.functions['my_filterfunction'] == my_filterfunction
-        assert J2Ninja.filters['my_filterfunction'] == my_filterfunction
+        assert J2Ninja.functions["my_filterfunction"] == my_filterfunction
+        assert J2Ninja.filters["my_filterfunction"] == my_filterfunction
